refactor(tests): mover base e testes do dominio chamado para app chamados
