+++ web/chamados/tests/test_models_chamado_contabilidade_nf_saida_validacao.py
@@ -1,5 +1,4 @@
 from django.core.exceptions import ValidationError
-
 from execucao.models import NF_SAIDA_ONLY_DIGITS_ERROR, Chamado
 from execucao.tests._base import ChamadoBaseTestCase
