+++ web/chamados/tests/test_service_validar_finalizacao.py
@@ -1,10 +1,10 @@
 from __future__ import annotations
 
-from chamados.services.finalizacao import validar_finalizacao
 from django.utils import timezone
-
 from execucao.models import Chamado
 from execucao.tests._base import ChamadoBaseTestCase
+
+from chamados.services.finalizacao import validar_finalizacao
 
 
 class ValidarFinalizacaoServiceTests(ChamadoBaseTestCase):
