+++ web/chamados/tests/test_views_chamado_execucao_get.py
@@ -2,7 +2,6 @@
 from __future__ import annotations
 
 from django.urls import reverse
-
 from execucao.models import Chamado
 
 from ._base import WebAuthBaseTestCase, grant_cap
